MIPS: baikal-t1: add configs for T-Platforms mITX boards
This is a minimal configuration to make mITX board up and running:
it includes device drivers for onboard chips only.

Unlike a generic config, defconfig doesn't include support for network filesystems
and block devices, as it's targeted for a local boot; loadable modules support
is disabled for now as well.

Device tree is based on bt1-gen and configs from Elpitech/baikal-t-linux-kernel repo.

Tested with BM1BT1-B board.

Signed-off-by: Nikita Radchenko <nradchenko@protonmail.com>